Modify html box classes to store objects.
The benefit of this is that an object can be bound to another at the beginning
of a page, then modified later. This removes the need for merging, and emphasises
that everything should be objects. It also makes it possible to modify an object
after it has been bound to the page, for instance.