Added basic userinfo class in PHP, created a base.php.
Base.php will hold common classes, like the box class
which doesn't have to be bound to HTML (which the userinfo
class demonstrates).

The idea of the userinfo get is that it will be used all
over the place. Both for the actually logged in user and
other users when doing admin-work. It will be extended to
support verification of user-data.

It is crucial at this time that browsers and the CSS support
dropdown-like menus. If we have to, we will extend the html
for specific browser-issues here, as it will be a key feature
for easy administration.

The news class went over a little change too, it now takes
an object as header1 and header2.