Added user class and multiuser classes
The multiuser classes is intended to be extended in other plugins, the
multiuserlist class is one way of extending it that will print a list
with all the users that were found with the userinfobox.

This is a good demonstration of how the database can be used with
callbacks efficiently.