Added some basic plugin support
This is untested and will most likely need a lot of work before it becomes
usable, and even more before it becomes good.

The basic idea is that plugins will be able to hook into certain core
classes on the fly, without modifying the classes for each plugin.
There is a slight problem with php not supporting multiple inheritance,
as I would like to have a "pluggable" class that supplies basic plugin
features not that diffrent from the box class. ...