Add login support in myuser class
This is will be improved over time, but the basic functionality is good.
The HTML code is temporary. It will be removed/fixed asap.

The idea is that $me will be used all over the place, but rarely printed.
You don't really want to print your own name everywhere, so it's not much
of a point. So when you DO print $me, it's to login.