Added basic event-awareness
As with everything, this will be extended along the way. The basic idea
is that we can have multiple BølerLAN's that would share the same genre,
but have diffrent event id's, and we could lend our bwreg installation
to diffrent lan parties who'd have a diffrent genre with their own
events.