Add storage backend for content and automatically add content to the page
This essentialy means that to add a page, you just add an entry to the
db with the name. The title has to be CamelCase to be reckognised as
a link automatically, but the wiki module can override the output.
[FrontPage frontpage] will for instance link to the magic FrontPage
but print frontpage. The first page is the genre name.