Add user: and news: wiki-links and add page-not-found warning.
The user: and news: links are only functional at the link-level, but will
hopefully work better when permissions are integrated and the userinfo
class is improved to support this, as it is intended to give an extensive
amount of information, based on the permissions of the logged in user.

This could be seating information for normal users, and payement status,
controls, message logs and whatnot for crew/admins.

The news: system will be easy enough. It will function exactly like the
content system wiki-wise, with uniqe names for the title as we've seen so
many other places. The bigger difference will be how they are built up and
such, which is easy enough with the news html class.

When a page isn't found, a message is added to the info4 box,
with a link to the FrontPage. This will need translation.

The idea is also to have this message include a "create this page"
link if the user has permission.

This also demonstrates how usefull it is to be able to just throw warnings
in a box, and still display the page, without thinking about if there are
other warnings there or not.