Auto refresh on contentsave, blockquote and content bottom margine to CSS
The autorefresh will help prevent users hitting refresh and re-posting.
The way it's used is a little silly at the moment, since the whole content
saving happens in get(), at which point it's too late to actually
set that tag (it's allready evaluated). The solution is to move the content
saving out of the get() method, which should've happened a while ago anyway.

... I smell an action handling system involving an array with objects.