Give users with BWReg2 permissions global permissions
This is mainly to ensure that they can maintain resources with "lost"
owners. (Ie: resources with only "rw" members). When they can create
resources and the resource editor has a saftey net for bad changes (ie:
people about to delete the last permission for a resource), the resource
control has more or less the functionality it needs.