Interface to change userinfo, improved userinfo
The userinfo class now takes care of what information it is allowed to print.
This is based in the 'private' flag in the user table. It allows users
to hide specific information about themself. This enables crew (for instance)
to display contact information, while regular user may only want to display
their name.

The newuser class is also extended to allow using it for updating an existing
user. This allows us to keep all rules regarding userinfo in one place. The
usage is trivial too.

We can now join and leave open groups easily, and sign up for "moderated"
groups. When the group-interface is created, the group signup process will
be refined.