Add user searching and improve userinfo display
The search is quite easy. The wording might need some refining though.
It is not possible to list all users like this. This is not meant as a way
for admins to list all users in any way, and more advanced searching for
admins will come when the ticket system appears. This is meant as a way
to locate a user for anyone who might want it.

This DOES have a weakness in that it will match info that is hidden, even
if it doesn't display it. That means you can search for "kristoffer", which
will find kristoffer, and show one "anonymous" entry. This allows you to
detrmine that a user named kristoffer is there.

Oh, the tables for userinfo display probably needs some CSS love and more...