Basic news displaying
Works quite well, but no interface to manage it yet.

To add news to a page, simply type: $NEWS:sname[,limit]$, where sname is
a uniqe identifier for that news type. ,limit is an optional number defining
the maximum ammount of news items to show.

It is perfectly safe to add a supersecret news to an otherwise fully public
page: The news are controled by their own ACL, not the one defined on the page.
This avoids the whole issue of a restricted users just adding $NEWS:secret$
to their own otherwise restricted page.

This will no doubt need adjustements.