Editing news and storing news
Not entierly perfect, still need to go over about two dousin FIXME's
and add a checkbox for wether news is EID specific or not. Also, some
checking of the content of the title so it's not a bunch of html...

Other than that, it works quite well. The news archive links show a
link to create new news if the users has write permission on any news
categories.

Also created the selectbox html class, and foption().