RSS support for news and minor tuneups
The header was an issue, I'm sure I can place the header() code at a smarter
place, but this works for now. The RSS stuff is pretty convenient, and pretty
crude. It could definitely use some love, but this works quite well considering
the state it's in.

This also includes some basic linking to news category control, but nothing
functional.

And a minor fix to the perm_list() so we take what permission
we're looking for (to only list writeable ACL's when we're doing
posting for instance). And class= support to htlinks.