Rough group administration
This works, but it feels messy. And there needs to be a more generic way
of controling options, without knowing about all of them.

This will need a lot of polish until I'm happy with it. I need user input
on exactly what they feel I can improve. General messy-ness is the only
specific thing I can think of, and lack of page-belonging.