More db work, currently broke the php code
InnoDB doesn't support auto_increment over multiple columns, so it had
to go. This means new content doesn't get a new contentid. Which we might
not need.

This finally introduces the majority of the foreign keys, though they
are not yet tuned. It also converts the final MyISAM table to InnoDB.
More DB work to come.